CRI-43 refactor: web-client naming
