+++ apps/web-client/src/app/app.tsx
@@ -1,4 +1,4 @@
-import './App.css';
+import './app.css';
 
 
 function App() {
