+++ apps/web-client/src/app/layouts/default-layout.tsx
