+++ apps/web-client/src/app/layouts/index.ts
@@ -1,0 +1,1 @@
+export { DefaultLayout } from './default-layout';
