+++ apps/web-client/src/app/providers/index.ts
@@ -1,0 +1,2 @@
+export { ProviderComposer } from './provider-composer';
+export { AppRouterProvider } from './app-router-provider';
