--- apps/web-client/src/application/layouts/index.ts
@@ -1,1 +1,0 @@
-export { DefaultLayout } from './DefaultLayout';
