--- apps/web-client/src/application/providers/index.ts
@@ -1,2 +1,0 @@
-export { ProviderComposer } from './ProviderComposer';
-export { AppRouterProvider } from './AppRouterProvider';
