+++ apps/web-client/src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
-import { ProviderComposer, AppRouterProvider } from '@/application/providers';
+import { ProviderComposer, AppRouterProvider } from '@/app/providers';
 
 import './index.css';
