+++ apps/web-client/src/pages/error/index.ts
@@ -1,1 +1,1 @@
-export { ErrorPage } from './ErrorPage';
+export { ErrorPage } from './error-page';
