+++ apps/web-client/src/pages/home/index.ts
@@ -1,1 +1,1 @@
-export { HomePage } from "./ui/HomePage.tsx";
+export { HomePage } from "./ui/home-page.tsx";
