+++ apps/web-client/src/pages/home/ui/home-page.tsx
@@ -1,13 +1,8 @@
-import { useNavigate } from 'react-router';
-
-import { WEB_ROUTES } from '@bc-arch-drafter/contracts';
-import { Button } from '@bc-arch-drafter/ui';
-
 import { BoardExample } from '@/features/board';
 
 export const HomePage = () => {
-  const navigate = useNavigate();
-  const toBaseForm = () => navigate(WEB_ROUTES.EXAMPLE);
+  // const navigate = useNavigate();
+  // const toBaseForm = () => navigate(WEB_ROUTES.EXAMPLE);
 
   return (
     <div className="flex flex-col items-center justify-center m-5 gap-5">
